+++ src/main/kotlin/it/czerwinski/intellij/wavefront/actions/ui/GenerateDiffuseIrradianceMapDialog.kt
@@ -16,17 +16,8 @@
 
 package it.czerwinski.intellij.wavefront.actions.ui
 
-import com.intellij.openapi.application.invokeLater
-import com.intellij.openapi.application.runReadAction
-import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.progress.progressSink
-import com.intellij.openapi.progress.withModalProgressIndicator
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.ProjectLocator
-import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.ui.ValidationInfo
-import com.intellij.openapi.vcs.vfs.VcsFileSystem
-import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.components.JBTextField
 import com.intellij.ui.dsl.builder.Panel
@@ -39,23 +30,18 @@ import java.awt.event.FocusListener
 import java.io.File
 import javax.swing.JComponent
 import javax.swing.JSlider
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.launch
 
-class GenerateDiffuseIrradianceMapDialog(private val project: Project) : DialogWrapper(project) {
+class GenerateDiffuseIrradianceMapDialog(project: Project) : BaseGenerateMapDialog(project) {
 
     private var samplesTextField: JBTextField? = null
     private var samplesSlider: JSlider? = null
     private var suffixTextField: JBTextField? = null
 
-    private var inputFiles: List<VirtualFile> = emptyList()
+    override val progressIndicatorTitle: String
+        get() = WavefrontObjBundle.message("action.GenerateDiffuseIrradianceMapAction.progress")
 
     init {
-        title = WavefrontObjBundle.message(
-            key = "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.title"
-        )
+        title = WavefrontObjBundle.message(key = "action.GenerateDiffuseIrradianceMapAction.title")
         init()
     }
 
@@ -66,9 +52,7 @@ class GenerateDiffuseIrradianceMapDialog(private val project: Project) : DialogW
 
     private fun Panel.samplesRow() {
         row(
-            label = WavefrontObjBundle.message(
-                key = "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.samples"
-            )
+            label = WavefrontObjBundle.message(key = "action.GenerateDiffuseIrradianceMapAction.samples")
         ) {
             samplesTextField = textField(
                 defaultValue = DEFAULT_SAMPLES,
@@ -104,9 +88,7 @@ class GenerateDiffuseIrradianceMapDialog(private val project: Project) : DialogW
 
     private fun Panel.suffixRow() {
         row(
-            label = WavefrontObjBundle.message(
-                key = "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.suffix"
-            )
+            label = WavefrontObjBundle.message(key = "action.GenerateDiffuseIrradianceMapAction.suffix")
         ) {
             suffixTextField = textField(
                 defaultValue = DEFAULT_SUFFIX,
@@ -115,66 +97,22 @@ class GenerateDiffuseIrradianceMapDialog(private val project: Project) : DialogW
         }
     }
 
-    fun show(inputFiles: List<VirtualFile>) {
-        this.inputFiles = inputFiles
-        show()
-    }
-
     override fun doValidate(): ValidationInfo? =
         if (suffixTextField?.text.isNullOrBlank()) {
-            val errorText = WavefrontObjBundle.message(
-                "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.suffix.errorEmpty"
-            )
+            val errorText = WavefrontObjBundle.message("action.GenerateDiffuseIrradianceMapAction.suffix.errorEmpty")
             ValidationInfo(errorText, suffixTextField)
         } else if (invalidCharacters.any { character -> character in suffixTextField?.text.orEmpty() }) {
-            val errorText = WavefrontObjBundle.message(
-                "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.suffix.errorChar"
-            )
+            val errorText = WavefrontObjBundle.message("action.GenerateDiffuseIrradianceMapAction.suffix.errorChar")
             ValidationInfo(errorText, suffixTextField)
         } else {
             super.doValidate()
         }
 
-    override fun doOKAction() {
+    override fun processFile(inputFile: VirtualFile): List<File> {
         val samples = samplesSlider?.value ?: DEFAULT_SAMPLES
         val outputSuffix = suffixTextField?.text ?: DEFAULT_SUFFIX
 
-        CoroutineScope(Job() + Dispatchers.Unconfined).launch {
-            @Suppress("UnstableApiUsage")
-            val outputFiles = withModalProgressIndicator(
-                project,
-                WavefrontObjBundle.message(
-                    "action.it.czerwinski.intellij.wavefront.actions.GenerateDiffuseIrradianceMapAction.progress"
-                )
-            ) {
-                progressSink?.fraction(0.0)
-                inputFiles.mapIndexed { index, inputFile ->
-                    progressSink?.text(inputFile.name)
-                    progressSink?.fraction(index.toDouble() / inputFiles.size.toDouble())
-                    DiffuseIrradianceMapGenerator.generate(inputFile, samples, outputSuffix)
-                }
-            }
-            invokeLater {
-                refreshFiles(outputFiles)
-            }
-        }
-
-        super.doOKAction()
-    }
-
-    private fun refreshFiles(outputFiles: List<File>) {
-        val virtualOutputFile = runReadAction {
-            val dirs = inputFiles.mapNotNull { it.parent }.distinct()
-            LocalFileSystem.getInstance().refreshFiles(dirs)
-            VcsFileSystem.getInstance().refresh(true)
-            LocalFileSystem.getInstance().findFileByIoFile(outputFiles.first())
-        }
-        if (virtualOutputFile != null) {
-            val project = ProjectLocator.getInstance().guessProjectForFile(virtualOutputFile)
-            if (project != null) {
-                FileEditorManager.getInstance(project).openFile(virtualOutputFile, true)
-            }
-        }
+        return listOf(DiffuseIrradianceMapGenerator.generate(inputFile, samples, outputSuffix))
     }
 
     companion object {